dbr quant: enable reference module support for torch.qint32
Summary:

This PR enables basic support for reference modules in DBR quant.
For now, the support is limited to:
1. modules that have reference versions defined only (no functions)
2. torch.qint32 dtype only

Currently, the reference module logic is enabled whenever dtype is
torch.qint32. This is done because this is needed the earliest for
the first use case. A future PR will support more dtypes and also
add the `is_reference` flag to the API.

Test plan:

```
python test/test_quantization.py TestQuantizeDBR.test_conv_int32_reference_model
```

TODO finish this

ghstack-source-id: 60df98be1935544e331160a291d4ca0287975d40
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73493