[ao][fx]refactor memoryless from user facing APIs
Summary: removed memoryless from user facing APIs and replaced with logic through the use of compute_dtype.
if compute_dtype isn't specified there are no changes to how this works, if compute_dtype is specified
then its used as the observer dtype (rather than the dtype which is treated as the activation_dtype).
Finally if compute_dtype is a quantized dtype and dtype=float (which is the designation for dynamic quantization we
came up with) then a memoryless observer is created.

Test Plan: python test/test_quantization.py

ghstack-source-id: 28900eef7aa30fff6238fe1b2429c97445028450
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73560